fixed bugs and create types
